+++ src/main/java/com/holub/model/cell/Neighborhood.java
@@ -1,7 +1,7 @@
 package com.holub.model.cell;
 
 import com.holub.life.Direction;
-import com.holub.life.Storable;
+import com.holub.tools.Storable;
 import com.holub.model.Point;
 import com.holub.tools.Observer;
