+++ src/main/java/com/holub/tools/Storable.java
@@ -1,4 +1,4 @@
-package com.holub.life;
+package com.holub.tools;
 
 import java.io.IOException;
 import java.io.InputStream;
